initial file restructuring
